DDPB-3181: Moved known common behat traits into Common dir
