+++ behat/tests/bootstrap/Common/BaseFeatureContext.php
@@ -3,9 +3,6 @@
 namespace DigidepsBehat\Common;
 
 use Behat\MinkExtension\Context\MinkContext;
-use DigidepsBehat\AuthenticationTrait;
-use DigidepsBehat\FormTrait;
-use DigidepsBehat\SiteNavigationTrait;
 
 class BaseFeatureContext extends MinkContext
 {
