+++ behat/tests/bootstrap/Common/RegionTrait.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DigidepsBehat;
+namespace DigidepsBehat\Common;
 
 use Behat\Gherkin\Node\TableNode;
