+++ behat/tests/bootstrap/Registration/RegistrationFeatureContext.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace DigidepsBehat\Registration;
+
+use DigidepsBehat\Common\BaseFeatureContext;
+
+class RegistrationFeatureContext extends BaseFeatureContext
+{
+    
+}
