+++ behat/tests/bootstrap/ReportManagement/ReportManagementFeatureContext.php
@@ -2,14 +2,11 @@
 
 namespace DigidepsBehat\ReportManagement;
 
-use DigidepsBehat\AuthenticationTrait;
 use DigidepsBehat\Common\BaseFeatureContext;
 use DigidepsBehat\Common\CourtOrderTrait;
-use DigidepsBehat\FormTrait;
-use DigidepsBehat\LinksTrait;
-use DigidepsBehat\RegionTrait;
-use DigidepsBehat\ReportTrait;
-use DigidepsBehat\SiteNavigationTrait;
+use DigidepsBehat\Common\LinksTrait;
+use DigidepsBehat\Common\RegionTrait;
+use DigidepsBehat\Common\ReportTrait;
 
 class ReportManagementFeatureContext extends BaseFeatureContext
 {
